requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
